Better-database

+++ training_gallery.py
@@ -1,38 +1,37 @@
 from __future__ import annotations
 
 import os
-import pathlib
 import pickle
 import sys
 
 import cv2
 import face_recognition
-from PyQt5 import uic
 from PyQt5.QtCore import Qt
 from PyQt5.QtGui import QPixmap
 from PyQt5.QtWidgets import QApplication
 from PyQt5.QtWidgets import QLabel
 from PyQt5.QtWidgets import QWidget
 
+from ui.ui_training_gallery import Ui_Images_qW
+
 CURRENT_FILE_PATH = os.path.dirname(os.path.abspath(__file__))
-UI_PATH = pathlib.Path(CURRENT_FILE_PATH, 'ui', 'training_gallery.ui')
-TRAINING_DATASET = pathlib.Path(CURRENT_FILE_PATH, 'resources', 'training_dataset')
-EXTRACTED_DATASET = pathlib.Path(CURRENT_FILE_PATH, 'resources', 'extracted_dataset')
-FACES_DAT = pathlib.Path(CURRENT_FILE_PATH, 'resources', 'faces.dat')
-ATTENDANCE = pathlib.Path(CURRENT_FILE_PATH, 'attendance')
+TRAINING_DATASET = os.path.join(CURRENT_FILE_PATH, 'resources', 'training_dataset')
+EXTRACTED_DATASET = os.path.join(CURRENT_FILE_PATH, 'resources', 'extracted_dataset')
+FACES_DAT = os.path.join(CURRENT_FILE_PATH, 'resources', 'faces.dat')
+ATTENDANCE = os.path.join(CURRENT_FILE_PATH, 'attendance')
 
 MAX_COLUMNS = 3
 
 
-class GalleryWindow(QWidget):
-    def __init__(self):
-        super().__init__()
+class GalleryWindow(QWidget, Ui_Images_qW):
+    def __init__(self, parent=None):
+        super().__init__(parent)
         self.CURRENT_FILE_PATH = CURRENT_FILE_PATH
         self.TRAINING_DATASET = TRAINING_DATASET
         self.EXTRACTED_DATASET = EXTRACTED_DATASET
         self.FACES_DAT = FACES_DAT
         self.ATTENDANCE = ATTENDANCE
-        self.ui = uic.loadUi(UI_PATH, self)
+        self.setupUi(self)
         self.init_ui()
         self.setup_button_events()
         self.get_images()
@@ -67,19 +66,19 @@ class GalleryWindow(QWidget):
             image_label = QLabel(self)
             image_label.setAlignment(Qt.AlignCenter)
             image_label.setPixmap(pixmap.scaledToWidth(200))
-            self.ui.Gallery_qGL.addWidget(image_label, row, col)
+            self.Gallery_qGL.addWidget(image_label, row, col)
 
     def delete_image(self):
         selected_label = self.get_selected_label()
         if selected_label is not None:
             filename = selected_label.property('filename') or self.images[selected_label.index][0]
             os.remove(os.path.join(self.image_directory, filename))
-            self.ui.Gallery_qGL.removeWidget(selected_label)
+            self.Gallery_qGL.removeWidget(selected_label)
             selected_label.deleteLater()
 
     def get_selected_label(self):
-        for i in range(self.ui.Gallery_qGL.count()):
-            widget = self.ui.Gallery_qGL.itemAt(i).widget()
+        for i in range(self.Gallery_qGL.count()):
+            widget = self.Gallery_qGL.itemAt(i).widget()
             if widget.property('selected'):
                 widget.index = i
                 return widget
@@ -124,33 +123,42 @@ class GalleryWindow(QWidget):
     def store_faces_with_names(self):
         faceClassifer = cv2.CascadeClassifier(f'{cv2.data.haarcascades}haarcascade_frontalface_default.xml')
 
-        for imgName in pathlib.Path(self.TRAINING_DATASET).glob('*.jpg'):
-            image = cv2.imread(str(imgName))
-            faces = faceClassifer.detectMultiScale(image, 1.1, 5)
-            name = imgName.stem
-            personPath = self.EXTRACTED_DATASET / name.split('_')[0]
+        for imgPath in os.scandir(TRAINING_DATASET):
+            if imgPath.is_file() and imgPath.name.endswith('.jpg'):
+                imgName = os.path.splitext(imgPath.name)[0]
+                image = cv2.imread(os.path.join(TRAINING_DATASET, imgPath.name))
+                faces = faceClassifer.detectMultiScale(image, 1.1, 5)
+                name = imgName
+                personPath = os.path.join(self.EXTRACTED_DATASET, name.split('_')[0])
 
-            if not personPath.exists():
-                personPath.mkdir(parents=True)
+                if not os.path.exists(personPath):
+                    os.makedirs(personPath)
 
-            for x, y, width, height in faces:
-                extracted_face = image[y:y + height, x:x + width]
-                resized_face = cv2.resize(extracted_face, (150, 150))
-                filename = f'{name}.jpg'
-                filepath = str(personPath / filename)
-                cv2.imwrite(filepath, resized_face)
+                for x, y, width, height in faces:
+                    extracted_face = image[y:y + height, x:x + width]
+                    resized_face = cv2.resize(extracted_face, (150, 150))
+                    filename = f'{name}.jpg'
+                    filepath = os.path.join(personPath, filename)
+                    cv2.imwrite(filepath, resized_face)
 
-            self.ui.message_qLB.setText('Treinamento feito com sucesso')
+        self.message_qLB.setText('Treinamento feito com sucesso')
 
     def train_faces(self):
-        directory = self.EXTRACTED_DATASET
+        directory = EXTRACTED_DATASET
         known_faces = []
         known_names = []
 
-        for namePath in directory.iterdir():
-            name = namePath.stem
-            for imagePath in namePath.glob('*.jpg'):
-                image = face_recognition.load_image_file(str(imagePath))
+        for name in os.listdir(directory):
+            namePath = os.path.join(directory, name)
+            if not os.path.isdir(namePath):
+                continue
+
+            for imageName in os.listdir(namePath):
+                imagePath = os.path.join(namePath, imageName)
+                if not os.path.isfile(imagePath):
+                    continue
+
+                image = face_recognition.load_image_file(imagePath)
                 face_locations = face_recognition.face_locations(image)
                 face_encodings = face_recognition.face_encodings(image, face_locations)
                 for encoding in face_encodings:
@@ -160,7 +168,7 @@ class GalleryWindow(QWidget):
         with open(self.FACES_DAT, 'wb') as f:
             pickle.dump((known_names, known_faces), f)
 
-        self.ui.message_qLB.setText('Faces extraidas e armazenadas com sucesso')
+        self.message_qLB.setText('Faces extraidas e armazenadas com sucesso')
 
     def train_model(self):
         self.store_faces_with_names()
